+++ src/mathlib-code.h
@@ -24,7 +24,7 @@ double division(double x, double y);
 
 double factorial(int x);
 
-double exponentiation(double x, double n);
+double exponentiation(double x, int n);
 
 double nthRoot(double x, double n);
